Create RawCode
This is very raw code but it does function. Those who are using this data themselves will need to provide their own Reddit Developer authentication information as seen at the top. This is only provided by reddit after being approved for a developer account. The rest is a matter of which parameters you would like to change depending on the target data you would like to find. Example - you can specify a different subreddit, you can specify x amount of posts to cycle through or even manipulate the threshold (Mine is 4>) for NLK occurrences you want to see for the final report.